requirements.txt
streamlit>=1.25
fpdf2>=2.7
qrcode[pil]>=7.4
Pillow>=10.0
